backend/requirements.txt
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
SQLAlchemy==2.0.20
PyJWT==2.8.0
python-dateutil==2.8.2